Current version of Psi+ is 1.1.33
It is based on:
* psi: 21bc537a
* patches: 92bc928
* plugins: 6787761
* resources: b261bbc

+++ iris/src/xmpp/sasl/scramsha1response.cpp
@@ -49,7 +49,7 @@ namespace XMPP {
 		QString salt = pattern.cap(2);
 		QString icount = pattern.cap(3);
 
-		size_t dkLen;
+		unsigned int dkLen;
 
 		QCA::Hash shaHash("sha1");
 		shaHash.update("", 0);
